+++ activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb
@@ -64,7 +64,7 @@ module ActiveRecord
           elsif conn_params && conn_params[:user] && error.message.include?(conn_params[:user])
             raise ActiveRecord::DatabaseConnectionError.username_error(conn_params[:user])
           elsif conn_params && conn_params[:host] && error.message.include?(conn_params[:host])
-            raise ActiveRecord::DatabaseConnectionError.hostname_error(conn_params[:host])
+            raise ActiveRecord::DatabaseConnectionError.hostname_error(conn_params[:host], conn_params[:port])
           else
             raise ActiveRecord::ConnectionNotEstablished, error.message
           end